--- src/Arius.Core/Features/ChunkHydrationStatusQuery/Models.cs
@@ -1,16 +1,0 @@
-using Arius.Core.Features.ListQuery;
-using Mediator;
-
-namespace Arius.Core.Features.ChunkHydrationStatusQuery;
-
-public enum ChunkHydrationStatus
-{
-    Unknown,
-    Available,
-    NeedsRehydration,
-    RehydrationPending,
-}
-
-public sealed record ChunkHydrationStatusQuery(IReadOnlyList<RepositoryFileEntry> Files) : IStreamQuery<ChunkHydrationStatusResult>;
-
-public sealed record ChunkHydrationStatusResult(string RelativePath, string? ContentHash, ChunkHydrationStatus Status);
